# [Inductor-CPU] Disable auto-tuning for templated int8 WoQ GEMM for small M to fix perf regression