### Summary

Described in #148494 - this PR fixes a regression (compared to the default Inductor-CPU behavior of not using max-autotune) for templated int8 WoQ GEMM (with BF16 activation) for small M dimension by disabling auto-tuning for small `M`, so that the ATen `_weight_int8pack_mm` kernel would be used.
The significance is next-token generation of LLMs.

Turning off auto-tuning for small `M` is a workaround. Ideally, we should improve the auto-tuning infra to prevent templated AVX512 GEMM for int8 WoQ being chosen if `_weight_int8pack_mm` would be faster E2E.
      
### Details

During auto-tuning, AVX512 GEMM micro-kernel is chosen for small `M`, but it's faster during auto-tuning, and performs worse E2E, which is expected as it can exploit cache locality for inputs while being called several times for the same inputs in a loop, but the same behavior isn't observed for its ATen counterpart `_weight_int8pack_mm`, which performs worse than it during auto-tuning but performs better E2E. However, it too would've benefited from better cache locality for inputs if it had been benchmarked for a longer time-period. Even so, the latency of the templated GEMM would still have been lower, even if we had benchmarked for more time.

|M | N | K | Templated GEMM latency during autotuning benchmarking | Templated GEMM latency E2E | `_weight_int8pack_mm` latency during autotuning benchmarking | `_weight_int8pack_mm` latency E2E | Ratio of E2E latency of templated GEMM over `_weight_int8pack_mm`|
|---|---|---|-------------------|----------------|----------------------|----------------|-----|
| 1|4096|4096|31.2 us |91.1 us |108.7 us | 76.07 us|1.19 |
|1|1024|4096| 16.1 us | 33.36 us | 52.9 us | 24.275 us |1.37 |
|1|14336|4096| 112.8 us | 274.16 us |335.3 us  |233.197 us| 1.17|
|1|4096|14336|128.1 us | 280.76 us| 330 us | 237.797 us| 1.18|
|1|4096|128256|1.642 ms|2.16 ms | 2.118ms|2.034 ms| 1.06 | 



### UTs

`python test/inductor/test_cpu_select_algorithm.py -v -k test_int8_woq_mm`


cc @voznesenskym @penguinwu @EikanWang @jgong5 @Guobing-Chen @XiaobingSuper @zhuhaozhe @blzheng @wenzhe-nrv @jiayisunx @ipiszy @yf225 @chenyang78 @kadeng @muchulee8 @amjames @chauhang @aakhundov