@sanchitintel 2025-03-04
@leslie-fang-intel, please advise if we should disable the templated GEMM kernel for int8 WoQ case on machines that don't support AMX ISA. Thanks!