@sanchitintel 2025-03-05
Hi Jiong, the idea is to choose ATen op `_weight_int8pack_mm` during auto-tuning because the ATen kernel performs better than the templated AVX512 GEMM end-to-end, but the templated AVX512 GEMM performs better during auto-tuning benchmarking.
AMX is too slow for small `M`, so the ATen kernel is chosen over it.
I guess another approach would've been to disable the AVX512 micro-kernel altogether for int8 WoQ with BF16 activation.


Some perf data (AMX vs `_weight_int8pack_mm`) -

```
AUTOTUNE _weight_int8pack_mm(1x4096, 4096x4096, 4096)
  _weight_int8pack_mm 0.1185 ms 100.0%
  cpp_packed_gemm_169 0.4729 ms 25.1%
SingleProcess AUTOTUNE benchmarking takes 0.2499 seconds and 3.8846 seconds precompiling for 2 choices
AUTOTUNE _weight_int8pack_mm(1x4096, 4096x4096, 4096)
  _weight_int8pack_mm 0.1058 ms 100.0%
  cpp_packed_gemm_173 0.5220 ms 20.3%
SingleProcess AUTOTUNE benchmarking takes 0.2503 seconds and 3.1777 seconds precompiling for 2 choices
AUTOTUNE _weight_int8pack_mm(1x4096, 16384x4096, 16384)
  _weight_int8pack_mm 0.3940 ms 100.0%
  cpp_packed_gemm_174 1.9697 ms 20.0%
SingleProcess AUTOTUNE benchmarking takes 0.2586 seconds and 3.7803 seconds precompiling for 2 choices
AUTOTUNE _weight_int8pack_mm(1x16384, 4096x16384, 4096)
  _weight_int8pack_mm 0.3885 ms 100.0%
  cpp_packed_gemm_175 1.7008 ms 22.8%
SingleProcess AUTOTUNE benchmarking takes 0.2565 seconds and 3.7806 seconds precompiling for 2 choices
AUTOTUNE _weight_int8pack_mm(1x4096, 50400x4096, 50400)
  _weight_int8pack_mm 0.9614 ms 100.0%
  cpp_packed_gemm_365 6.1783 ms 15.6%
```


Thanks!