@leslie-fang-intel 2025-03-05
> One reason for this behavior could be different cache behavior with respect to input data (for auto-tuning benchmarking vs. E2E model run). A secondary reason could be that the ATen kernel is generic but there are many codegened GEMMs created, which are not as i-cache friendly as the ATen kernel (since there's only one _weight_int8pack_mm but many codegened kernels), and the i-cache misses are not amortized by the speedup the codegened kernels bring in this case.

How much perf gap did we see for a specific GEMM between ATEN kernel and AVX512 Micro GEMM? For both the benchmark data and model runtime data.