@sanchitintel 2025-03-05
Hi @leslie-fang-intel, the rationale for issue can be proved simply by benchmarking with int8 WoQ enabled & disabled for next-token generation of an LLM. Only linear layers are quantized, so GEMMs are the only differentiating factor.

The data you asked for would help in debugging the issue, but why not follow the standard operating procedure of fixing a regression?

Thanks!