@sanchitintel 2025-03-05
Hi @jgong5, I modified this PR to disable auto-tuning for int8 WoQ GEMM case for `M < 32` (templated GEMM is slower for M < 32, irrespective of whether AVX512 or AMX micro-kernel used). For `M >= 32`, the AMX micro-kernel would be used on machines that support the AMX ISA.