@jgong5 2025-03-06
> > why the template GEMM with AVX512 is slower than AMX counterpart
> 
> It's faster
> 
> @jgong5, the idea behind this PR is to choose ATen op `_weight_int8pack_mm` for small `M` during auto-tuning because the ATen kernel performs better than the templated AVX512 GEMM end-to-end, although the templated AVX512 GEMM performs better during auto-tuning benchmarking. We had discussed this issue last year as well.
> 
> In this PR, I deliberately (re-)enabled AMX GEMM even for small `M`. because it's quite slow, so the ATen kernel (`_weight_int8pack_mm`) would be chosen over it during auto-tuning.
> 
> I guess another approach would've been to disable the AVX512 GEMM micro-kernel altogether for int8 WoQ with BF16 activation, so that it wouldn't be chosen even on machines that lack AMX support (I haven't collected perf data on such machines, though).
> 
> Thanks!

I understand the problem of differed performance you saw between ATen kernel and templated AVX512 GEMM kernel during auto-tuning and end-to-end runs. It is definitively something we need to fix. But I am concerned about the way you address the problem. It is more like a temporary workaround and this workaround promotes an even worse kernel (AMX GEMM kernel) for small "M". If we force the compilation to use the template GEMMs, we would have a regression with your fix. Can you share perf numbers as I requested for us to know the problems better?