@sanchitintel 2025-03-06
> Can you share perf numbers as I requested for us to know the problems better?

The ATen kernel `_weight_int8pack_mm` performs better E2E than during auto-tuning benchmarking!
And its E2E performance is better than that of templated GEMM with AVX512 micro-kernel!

|M | N | K | Templated GEMM latency during autotuning benchmarking | Templated GEMM latency E2E | `_weight_int8pack_mm` latency during autotuning benchmarking | `_weight_int8pack_mm` latency E2E | Ratio of E2E latency of templated GEMM over `_weight_int8pack_mm`|
|---|---|---|-------------------|----------------|----------------------|----------------|-----|
| 1|4096|4096|31.2 us |91.1 us |108.7 us | 76.07 us|1.19 |
|1|1024|4096| 16.1 us | 33.36 us | 52.9 us | 24.275 us |1.37 |
|1|14336|4096| 112.8 us | 274.16 us |335.3 us  |233.197 us| 1.17|
|1|4096|14336|128.1 us | 280.76 us| 330 us | 237.797 us| 1.18|
|1|4096|128256|1.642 ms|2.16 ms | 2.118ms|2.034 ms| 1.06 |



> But I am concerned about the way you address the problem. It is more like a temporary workaround and this workaround promotes an even worse kernel (AMX GEMM kernel) for small "M". 

@jgong5, yes it's a workaround. `_weight_int8pack_mm` would be used instead of templated GEMM for small `M`.
Also, y'day, I disabled auto-tuning for small `M` altogether for int8 WoQ case.

> If we force the compilation to use the template GEMMs, we would have a regression with your fix. 

Sorry, is there any plan to force use of templated GEMMs?

Thanks!