@sanchitintel 2025-03-07
> User can set `max_autotune_gemm_backends="CPP"` to force use it.

Thanks, @leslie-fang-intel! The revision wouldn't run into that problem but I'm currently looking into improving the auto-tuning implementation to avoid using this workaround.